+++ csharp/test/Microsoft.ML.OnnxRuntime.Tests/InferenceTest.cs
@@ -231,13 +231,13 @@ namespace Microsoft.ML.OnnxRuntime.Tests
 
 #if USE_TENSORRT
         [Fact]
-        private void validateTensorRTProviderOptions()
+        private void TestTensorRTProviderOptions()
         {
             string modelPath = Path.Combine(Directory.GetCurrentDirectory(), "squeezenet.onnx");
             string calTablPath = Path.Combine(Directory.GetCurrentDirectory(), "squeezenet_calibration.flatbuffers");
             //Environment.SetEnvironmentVariable("ORT_TENSORRT_ENGINE_CACHE_ENABLE", "1");
 
-            SessionOptions.OrtTensorRTProviderOptions trt_options = SessionOptions.GetDefaultTensorRTProviderOptions();
+            OrtTensorRTProviderOptions trt_options = ProviderOptions.GetDefaultTensorRTProviderOptions();
             trt_options.device_id = 0;
             trt_options.trt_int8_calibration_table_name = calTablPath;
             trt_options.has_trt_options = 1;
